Begin converting `TTMTask` class to Swift
This is a big class so it would be a lot of work to try to convert it
all at once.  By converting the Objective-C class to a base class and
then deriving a Swift class from it, we can convert groups of methods
incrementally while allowing the Objective-C class to continue to handle
the rest.

This first increment converts all of the task recurrence methods to
Swift, which also eliminates the need for the Objective-C `NSDate`
extension methods.